+++ src/lib/agent/plugins/argentx/constant/contract.ts
@@ -1,0 +1,5 @@
+export const argentx_classhash =
+  '0x036078334509b514626504edc9fb252328d1a240e4e948bef8d0c08dff45927f';
+
+export const DEFAULT_GUARDIAN = '0x0';
+export const DEFAULT_NONCE = '0x0';
