+++ src/lib/agent/plugins/braavos/constant/contract.ts
@@ -1,0 +1,10 @@
+export const braavos_proxy_classhash =
+  '0x03131fa018d520a037686ce3efddeab8f28895662f019ca3ca18a626650f7d1e';
+export const braavos_intiial_classhash = 
+  '0x5aa23d5bb71ddaa783da7ea79d405315bafa7cf0387a74f4593578c3e9e6570';
+export const braavos_account_classhash = 
+  '0x2c2b8f559e1221468140ad7b2352b1a5be32660d0bf1a3ae3a054a4ec5254e4';
+
+export const DEFAULT_GUARDIAN = '0x0';
+export const DEFAULT_NONCE = '0x0';
+
