+++ src/lib/agent/plugins/core/contract/constants/contract.ts
@@ -1,14 +1,2 @@
-export const argentx_classhash =
-  '0x036078334509b514626504edc9fb252328d1a240e4e948bef8d0c08dff45927f';
-export const oz_classhash =
-  '0x061dac032f228abef9c6626f995015233097ae253a7f72d68552db02f2971b8f';
-export const braavos_proxy_classhash =
-  '0x03131fa018d520a037686ce3efddeab8f28895662f019ca3ca18a626650f7d1e';
-export const braavos_intiial_classhash = 
-  '0x5aa23d5bb71ddaa783da7ea79d405315bafa7cf0387a74f4593578c3e9e6570';
-export const braavos_account_classhash = 
-  '0x2c2b8f559e1221468140ad7b2352b1a5be32660d0bf1a3ae3a054a4ec5254e4';
-export const okx_classhash =
-  '0x01c0bb51e2ce73dc007601a1e7725453627254016c28f118251a71bbb0507fcb';
 export const DEFAULT_GUARDIAN = '0x0';
 export const DEFAULT_NONCE = '0x0';
