+++ src/lib/agent/plugins/okx/constant/contract.ts
@@ -1,0 +1,6 @@
+export const okx_classhash =
+  '0x04d07e40e93398ed3c76981e72dd1fd22557a78ce36c0515f679e27f0bb5bc5f';
+  
+export const DEFAULT_GUARDIAN = '0x0';
+export const DEFAULT_NONCE = '0x0';
+
