+++ src/lib/agent/plugins/openzeppelin/constant/contract.ts
@@ -1,0 +1,5 @@
+export const oz_classhash =
+  '0x061dac032f228abef9c6626f995015233097ae253a7f72d68552db02f2971b8f';
+
+export const DEFAULT_GUARDIAN = '0x0';
+export const DEFAULT_NONCE = '0x0';
